+++ src/torchjd/linalg/__init__.py
@@ -1,0 +1,8 @@
+"""
+This module provides type annotation classes representing tensors with specific structural
+properties.
+"""
+
+from torchjd._linalg._matrix import Matrix, PSDMatrix
+
+__all__ = ["Matrix", "PSDMatrix"]